Fix object detection pipeline
Oject detection is run in a distinct job.
The pipeline was broken, as the job did not have access to the
DB Image instance necessary to perform the detection (each job run in a
distinct SQL transaction).
This restores the functionality, by running OCR image insight import
and object detection sequentially in the same job. The 1st step is in
a transaction that ends before the object detection starts, which
means the DB image instance is available.