fix: fix insight import mechanism
When a prediction based on a more recent image is available, we don't
transform this prediction into an insight, we keep the old insight.
This is problematic as the insight based on the newer image may make it
automatically processable (if the image is selected or if it's not too
old compared to most recent images).
As knowing which insight to keep/delete and which candidate to import
is tricky (many subtle bugs may occur), I chose here a more radical
approach: we regenerate all insights for a product from predictions, we
only keep annotated insights in DB.

I also added a better sorting of candidates:
- we now set a higher priority on candidates that are automatically
processable (this field may be set after the ProductInsight creation,
that's why we cannot sort according to this field in `sort_predictions`)
- we now set a higher priority on candidates that have a non null
predictor field