feat: consider annotation=-1 as a vote (#908)
So that this mechanism does not have to be implemented client-side.
The voting mechanism has also been updated: before, all classes
(0, 1, -1) could be considered as a valid value. Now, -1 is exclusively
used for users to ignore insights, so it's not taken into account when
computing vote counts (only 0 and 1 are considered).
Another bug was fixed: when vote counts for 0 and 1 are both >= 2,
we save the insight as invalid (-1) instead of saving it as negative
(0).