fix: remove automatic parameter in InsightImporter
This parameter is a legacy parameter that does not make sense anymore to
keep: it was used so that insights generated from OCR dump could not be
applied automatically, as it could come from old images. Now we have
a robust system to know whether to apply the insight automatically, we
don't need this parameter anymore.