feat: introduce prediction deletion
So far, we didn't delete any old prediction, and this had to be done
manually everytime we introduced a breaking change in every prediction
type.
To ease the process, we add a new `predictor_version` to Prediction and
ProductInsight tables. Everytime we introduce a breaking change, we
increase the `predictor_version`. During import, we delete predictions
associated with other `prediction_version`s.