fix: improve worker task dispatch
before, we had 2 high priority workers (that listened to queue high-1
and high-2 respectively), and 2 workers that listened both to one
high priority queue (high-3 or high-4) and to low priority queue.

It's beneficial that all workers listen to low priority queue (but with
a lower priority), so that low priority jobs get completed faster when
there is no high priority tasks.