fix: take into account product type changes
When the `product_type` field is changed, it means we switch from one
flavor to another one. So far, such a change had no effect on Robotoff
(with the exception that the insights and predictions were deleted the
night after by the scheduled maintenance task).

Now, we delete elements in all tables that are related to the product
and reimport the images with the new product type.
More specifically, we delete for the product:
- all logo annotations, from DB and Elasticsearch
- all image predictions from DB
- all images from DB
- all predictions from DB
- all non-annotated insights from DB

We also schedule:
- the reimport of all images with the new product type
- the insight update for the product (including the category
prediction).