feat: upgrade logo detector model (#1781)
## feat: upgrade universal-logo-detector model to Yolo architecture

- use newly trained universal-logo-detector model, using Yolov8 arch
- remove completely support for Tensorflow object detection models

## fix: use the same padding color as in ultralytics

In ultralytics, color (114, 114, 114) is used as a filling color, not
black

## fix(object-detection): fix the NMS values

The values were not the correct ones for all Ultralytics models.
NMS threshold is 0.7 by default, and not 0.45. It resulted in
discrepancies between inference using Ultralytics package and Robotoff.

## feat(object-detection): add albumentations preprocessing (optional)

Allow to preprocess images with albumentations instead of custom
preprocessing. It should be faster, and is an exact copy of what is
done on Ultralytics, contrary to the custom implementation
(Pillow-based).

## feat(object-detection): use np array as input for all object det
models